pyproject.toml
[tool.poetry]
name = "Credit KT"
version = "0.1.0"
description = "KT"
authors = ["Sai Thiha Zaw <sai.t.zaw@frontiir.net>"]
license = "BSD"

[tool.poetry.dependencies]
python = "3.7.12"
pandas = "^1.3.5"
jupyter = "^1.0.0"
jupyterlab = "^3.2.5"
scipy = "^1.7.3"
matplotlib = "^3.5.1"
seaborn = "^0.11.2"
python-dotenv = "^0.19.2"
psycopg2-binary = "^2.9.3"
SQLAlchemy = "^1.4.29"
pendulum = "^2.1.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
